add bin_test tag
The bin_test tag should help others avoid the foot-gun of trying to run
tests that require a Nomad binary against the wrong binary.
